2. moving date constants, initializing date flatpickr
